Updated ColorGFX pause to signal an interrupt instead of calling join().
Calling join() blocks the thread that calls it. In this case, the UI
thread was being blocked which is bad and has the potential in resulting
in an ANR. Interrupt is now used to signal the end of a thread's life.
It does not block the thread that calls it.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>