Created a DBUpgradeTask which creates a worker thread for DDL execution.
Since all database setup and upgrades occur via DDL statements stored in
local files, complex setups may take a little bit longer. This will
create a worker thread for these setup and upgrades ops. The user will
see a notification on the screen stating that an upgrade is in progress
and an indeterminate progress indicator is displayed.

In addition to these database updates, the SplashActivity will now wait
for any necessary updates to occur before displaying the MainActivity
(which will require a completely setup database).

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>